refactor: BuildingRegisterQueryService 구현(#100)
